fix: drop redundant int64 conversion (lint) + address review comments (#116)
- prometheus.go: geoIPBlocked is now atomic.Int64, so int64(geo) was an
  unnecessary conversion the unconvert linter rejected (CI test job).
- bench_test.go: ResetTimer before RunParallel so benchMiddleware setup is
  excluded, matching the other benchmarks.
- caddywaf.go / handler.go: corrected comments for the atomic counters vs the
  lock-guarded ruleHitsByPhase map, and for incrementGeoIPRequestsMetric (it
  bumps geoip_blocked only when blocked). Per Copilot review on #190.

+++ prometheus.go
@@ -73,7 +73,7 @@ func (m *Middleware) renderPrometheus() string {
 	counter("total_requests", "Total requests processed (process-local).", total)
 	counter("blocked_requests", "Total requests blocked (process-local).", blocked)
 	counter("allowed_requests", "Total requests allowed (process-local).", allowed)
-	counter("geoip_blocked", "Requests blocked by country/ASN (process-local).", int64(geo))
+	counter("geoip_blocked", "Requests blocked by country/ASN (process-local).", geo)
 	counter("ip_blacklist_hits", "IP blacklist hits (process-local).", ipHits)
 	counter("dns_blacklist_hits", "DNS blacklist hits (process-local).", dnsHits)
 	counter("rate_limiter_requests", "Requests counted by the rate limiter (process-local).", rlReq)
